requirements.txt
Django
djangorestframework
djongo
pandas
requests
python-dotenv
streamlit
